requirements.txt
eventlet==0.23.0
requests==2.18.4
discord.py == 0.16.12
pyparsing ==


